==> 02.Rmd <==
# `r {load('_init.RData');kopi();tdis}` {#dis}

#### Abstract {-}

(ref:abs-dis)

#### Keywords {-}

(ref:key-dis)

***

The classification of cultural products according to genre systems has been treated both as a top-down institutional process conditioning the development of entire markets [@Lena2008Classification; @Brook1994Symphonie] and as "nothing but" an emergent, bottom-up process of individual acts of labeling and interpretive judgement [@Lamont2010How; @Hitters2010Tune]. A middle range approach requires understanding how actors distributed across the roles that constitute structure, the top down perspective, make local judgements that either do or do not have global consequences. The idea of genre as a cognitive institution that is diffused across a wide variety of actors gives the appearance of a cultural integrator, something given in the environment of all actors and not controlled by any of them. Players on fields of cultural production regardless of their role or status are, like price takers in a market, cultural name takers. Once genre labels are established they are immutable and given as terms in the language. The content to which they refer, however, is historically variable and debatable.

The explanation of how genre labels and the content to which they refer develops over time requires a mapping of the structure of genre labels at particular moments. In this study I estimate a mapping of the cognitive associations among genre categories of scholarship according to one interested role position in the scholarly field of cultural production, the archivist. Archives play a constitutional role for scholarly fields, forming a necessary infrastructure for scholars' work, but scholars are not the only interests served by archives. Students and educators as well as the public also draw on knowledge resources contained in the archive. To make this knowledge accessible archivists use genre classifications to organize content. The classifications meet the practical needs of serving both the producers and consumers of the content. The granularity of the classification system that would be useful to producers is, however, much finer than what is useful for consumers (note that the same item may either be factor of production or a product). The bias of the archive skews toward the courser grained consumer classification system in part because the needs of producers are highly idiosyncratic, shifting, and difficult to predict.

Producers tend to dismiss the consumers' ways of thinking about their field, but there is one situation in which producers are reduced to consumers; in scholarly professions the coarse version of classification is a communicative infrastructure for foreign correspondence. These coarse categories are what we call disciplines. Disciplines are nothing but a genre like classification system; they are merely integrative of global transactions among scholars at a distance and between scholarship as a whole and its macroscopic interlocutors, namely education, grant making, and university administration. Thus when scholars view content situated outside of their local epicenters of research, they necessarily shift from fine to coarse grained judgements. This switch on the tracks may be jolting, as it connotes what they feel to be a demotion of their usual status.

```{r fractal}
gv<-DiagrammeR::grViz('
  graph "" {
    graph [center=true]
    
    tooltip=" "
    node [tooltip=" "]
    edge [tooltip=" "]
    
    node [shape=plaintext,fontname=Courier,fontsize=12]
    m [label=Method]
    ql [label=Qualt]
    qn [label=Quant]
    ql1 [label=Qualt]
    qn1 [label=Quant]
    ql2 [label=Qualt]
    qn2 [label=Quant]
    m--ql
    m--qn
    ql--ql1
    ql--qn1
    qn--ql2
    qn--qn2
    ql1--"Very\nQualt" [style=dotted]
    qn1--"Somewhat\nQualt" [style=dotted]
    qn2--"Very\nQuant" [style=dotted]
    ql2--"Somewhat\nQuant" [style=dotted]
  }
  ')
htmlwidgets::saveWidget(gv,file='/home/plagiat/portfolio/exh/fractal.html')
webshot::webshot('/home/plagiat/portfolio/exh/fractal.html',paste0('img/fractal.',ec('png,pdf')),selector = '#graph0')
```

```{r fractalp,include=T,fig.cap='Fractal Evaluation and Commensurated Scale. Adapted from Abbott (2011:11).',fig.width=3,fig.asp=500/610,out.width=NULL}
knitr::include_graphics('img/fractal.png',auto_pdf = F)
```

Facing inwardly, producers deploy value standards rather than classifications. In their own field, scholars grant themselves moral and aesthetic authority to judge the quality of content. In communicating outside of their field, where they have no jurisdiction, assessment of cultural items becomes value neutral, a mere matter of taste. Such internal values may even be simple dualisms compared to the more complicated disciplinary categories, but rather than creating simple genre-like schemes they create scales of gradient evaluation. @Abbott2001Chaos shows that these scales, for instance from idiographic to nomathetic, are established by the accumulation of simple judgments by separate individuals in the direct and often pairwise comparisons of discrete works or authors. This pattern of evaluation is chaotic and produces a fractal organization of content with only approximate correspondence among judges. Though Abbott [-@Abbott2001Chaos\:14] suggests that these fractals are irreducible to simple one or two-dimensional scales, it is likely that the chaotic outcomes are collapsed into a reified commensuration standard when facilities and rewards are distributed [@Espeland1998Commensuration].

Disciplines are much more like genres than the constellation of fields and subfields that scholars may be accustomed to thinking with in the course of their work. Genres function to the extent that they are globally recognizable, that is, ubiquitous. Ubiquity means that all members of a field regardless of status are able to learn enough about the genre to make accurate classifications of cultural objects. This does not mean that everyone will always necessarily be aware of all genre labels, or that their classifications will always be accurate, but it does mean that they will be able to learn the rules when it becomes necessary. Subfield categories are not ubiquitous in this way; subfield labels only become genre-like by restricting the reference group to a more esoteric level. Whereas all members of a university, for instance, may easily become aware of the discipline of sociology, written as it may be on a building, only disciples of sociology are likely to become aware of the subfield of conversation analysis. And while disciples may learn the label, it takes some real experience with the content of the discipline, if not necessarily the subfield, to understand that conversation analysis is more qualitative than population research and more quantitative than ethnography.

In this way the classification schemes are nested in a functional hierarchy that shifts according to which role relationships are activated in the field. The more fine grained distinctions among genres are derivable from the collapsed fractal patterns. The highest level of granularity one is able to discuss may be a status signal of where they belong within a disciplinary stratification system. Furthermore the interactional capacities of disciples are nested. Denizens of more interior spaces in a subdiscipline, the spaces of fractal evaluation, are always able to operate successfully in more granular interactions, but the reverse is not true.


```{r role-schema,include=T}
eg('Accessibility,Schema,Ego,Alter
esoteric,fractal evaluation,scholar,scholar
.,genre class,scholar,student
..,,,publisher
...,,,archivist
....,,,department
ubiquitous,discipline class,scholar,administrator
,,,public'
)[[1]] %>%
  sg(tit='Classificatory Schema by Role Relationship',new.col.align = 'llll')
```

Everyone starts at the more ubiquitous level. It is only through education or professional socialization that newcomers can learn genre classifications, and most members of the public will not even be aware of the disciplinary categories of academic knowledge. The genre system sits in a more optimized relationship between descriptiveness and stability. The disciplines only change a glacial pace, while the genres are stable enough to maintain relevance for entire generations of scholars.  The fractal patterns are so granular that they may not even survive in the memory of a particular person over the years. In its constant interaction with the chaos of actual knowledge production, the genre system provides a more stable medium that is open to change but not so open as to risk dissolution. Scholars may complain that their work is not adequately represented by the genre system, and if enough make the same argument, the genre system may update to allow a wider audience to recognize the novel work. This is not to say that the work need actually be novel. As Abbott observes, each generation of scholars tends to cycle through forgetting and then reinventing the knowledge it had just cast aside [-@Abbott2001Chaos\:24].

The level of disciplines reformats the interaction of fractal evaluation and genre codification, just at a higher resolution that is oriented to the external relationships of the university. Where knowledges are separated by great cultural if not necessarily semantic distances, such as for instance the divide between biology and sociology, scholars in distant disciplines may not have much better knowledge about each other than a random member of the public. Disciplines emerge at a societal scale in a way that lets the university enter exchange relationships with government and the economy.


This definition of discipline is isomorphic with that of genre, in that disciplines are another form of categorical decision making, just at a higher level. Disciplines are supergenres that reflect the nesting of categories. In music Lena and Petersen [-@Lena2008Classification\:698] have followed Ennis [-@Ennis1992seventh] in referring to supergenres like rock-n-roll or rhythm and blues "streams", which are genres linked in time by any of a number of contiguities, e.g. sound, scene, artist, or market. What may distinguish disciplines in a more serious way is the institutional underpinning of the category. If genres are formed by the categorical thinking of disciples, disciplines are formed by the categorical thinking of patrons, the powerful actors like university administrators, state legislators, and grant makers who manage the economic foundation of scholarly careers, as well as the elite disciples who interface with these stakeholders. Where genre labels are inscribed onto cultural products, texts, discipline labels are inscribed onto cultural infrastructure like funding lines, buildings, and personnel themselves.

The uses of genre labels are themselves systematically organized by discipline, and a disciple who is adequately trained will know the correct ways to use genre labels in her local context. To use genre as a disciplinary convention means to first identify your location in the disciplinary field, and then to accept the limitations on scope by excluding knowledges and relationships that are extradisciplinary, that is, irrelevant. Disciplinary structure reduces the true cultural complexity of the entire genre system to a restricted form of local relevance, which in turn allows humble knowledge workers to engage upon a set of shared assumptions. Such simplicity begets new complexity as disciples spin out the consequences of their local use of genre.

The genres are also economic principles helping to organize supply and demand within internal scholarly markets for cultural products. There actors see genres among the borders between economic, social, and cultural uses, as a market category helping them acquire or produce content, as a card in proximate games of prestige, or as something to taste, to consume and enjoy directly. Publishers and journal editors leverage genres to help producers and consumers find them, as proclaiming a venue's genre identity is the cheapest way to advertise. Still, journals must also contend with the fractal evaluation of producers, and this may lead to creative positioning within the genre system if not usually the disciplinary system. As a way to signal that they are not merely generic, which is to say boring, a journal may position itself between rather than squarely within genres. This kind of recombinant innovation maintains the stability of both the genre and the discipline systems while at the same time opening up space for the fractal distinctions to work themselves out in creative ways [@Vedres2010Structural;@Gitlin2000Triumph].

What I seek to accomplish in the remainder of this study is to understand how the global disciplinary structure is reflected in the indexing behavior of archivists as they attempt to locate journals within a given classification scheme. Archivists have an important role is distributing the classification system and regulating it across a large population of knowledge seekers. While they do not author the discipline categories they do provide the an infrastructure that facilitates their taken-for-grantedness. In attempting to understand how a particular item of cultural content fits into a classification system, their only degree of freedom is to apply multiple labels to the item. This turns a binary distinction, is this item discipline-relevant or not, into something more nuanced, this item is partly relevant and partly not.

I use the JSTOR archive as a case of the work of archivists to index cultural material in a way that is useful to consumers. This requires making classification decisions about underlying items. In this case they make decisions about which discipline, which they call subjects, particular journals belong to. And to make every journal fit, they may apply multiple labels in an attempt to cover the content of the journal in a parsimonious way. By observing how the patterns of multiple classification are organized we can see how the disciplinary classification system is adapted to meet the requirements of more granular levels. Rather than marking boundaries of strict relevance, what we will find are regions of proximity within which partial relevance is established. This allows scholars to be innovative without losing their grounding in the discipline system. And for journals who have an interest in stable reproduction of scholarly activity, it allows publishers and editors to position themselves as having jurisdiction over particular kinds of innovation.


## JSTOR Journals {#kd-dq1}

```{r master2jstorm}
f<-'d/q/jstorm.RData'
if(file.exists(f)) {
  load(f)
} else {
  jstorm<-master2jstorm.f()
  save(jstorm,file=f)
}
rm(f)
```

```{r jstorm2jclu}
f<-'d/q/jclu.RData'
if(file.exists(f)){
  load(f)
} else {
  jclu<-jstorm2jclu.f(jstorm,shallow=T) # TODO consider two passes, one identifying sups, the second removing them
  save(jclu,file=f)
}
rm(f)
```

JSTOR provides a title list of their journal coverage [@JSTOR2018Title]. The coverage of journals in the archive is very complete for those journals chosen for the database. As of this writing JSTOR contained `r nn(jclu$n$t)` different journal titles and `r nn(jclu$n$j)` journals from `r nn(jclu$n$p)` different publishers. The different journal counts are due to some journals changing titles at least once.^[To avoid over counting, title histories are collapsed into their most recent record, meaning all subsequent counts are out of `r nn(jclu$n$j)`. Even though one might expect disciplinary identity to change over time, JSTOR discipline labels do not vary within title histories. `r {dN<-jstorm[title_id==title_history,sapply(discipline,length)]==0;if(any(dN)) dS<-sum(dN);paste0(nn(dS,cap=T),' journal',if(dS>1) 's','--',jstorm[title_id==title_history][dN,nl(publication_title)],'--lacked any discipline labels and ',if(dS>1) 'are' else 'is',' excluded from tabulations.')}`] The JSTOR coding contains `r nn(jclu$n$l)` subject labels. These labels refer to `r nn(jclu$n$s)` superdisciplines under which may be found `r nn(jclu$n$d)` disciplines.

Most journals are given more than one discipline label, and the superdisciplines are not marked as such in the database  creating some redundancy. For instance, a journal labeled as "Sociology" will also be labeled as "Social Sciences". Most academics will be familiar with whether a label is for a superdiscipline or a subdiscipline, yet for outsiders or for skeptical insiders, the only clue is in the frequency with which a label is applied. Counting labels, however, does not unambiguously place a journal in one discipline or another because journals may bear multiple labels, even multiple superdiscipline labels.

## Network Mode Projection

To assess the size of the disciplines and to disentangle their hierarchies it will be helpful to have a mutually exclusive labeling scheme that draws on the JSTOR curators' judgement while simplifying it. I rely on network methods to accomplish this labeling in a data driven and reproducible way. In a network representation of journal discipline labels, two journals may be said to be be related if they carry the same label. In network terms this can be represented as a bipartite or bimodal network. In a bimodal network there are two types (modes) of nodes, a journal and a label, and ties can only be recorded between, not within, these modes. So journals are not tied directly to other journals and labels are not tied directly to other labels.

Given any bimodal network, one may translate or project it into either of two unimodal forms. In a single mode or unimodal projection of a bimodal network there is only one type of node, in my case either a journal or a label, but not both. The omitted type is instead represented as a set of ties among the included type. Though the bimodal network is a more elegant representation, it is technically necessary to project it into one of its two bimodal forms to leverage network methods that are designed with unimodal data in mind.

Using the list of subjects associated with each journal in the JSTOR title list, I construct the bimodal *journal-label* network with journals in the first mode bearing ties to discipline labels in the second mode. I then project the bimodal network into two unimodal networks, one where journals are connected by ties equal to the number of discipline labels they have in common, and another where labels are tied by the number of articles carrying both labels. Call each of these unimodal networks, the (*journal-label-journal*) journal network and (*label-journal-label*) label network, a facet of the original bimodal network.

`r {wrk<-jstorm[title_history==title_id,sapply(discipline,length)];wrk %<>% .[.!=0];tab<-table(wrk)}`
Figure \@ref(fig:mod-proj) illustrates the effects of network mode projection on a random sample of 300 edges from the full JSTOR title list network. The first panel illustrates the bimodal network where journals are yellow dots and labels are blue dots. As an artifact of sampling, most journals here are shown tied to only one label. In fact this is never the case in the full network; as each journal has at least one discipline and one superdiscipline label the minimum number of labels is two, which is the median case accounting for `r (prop.table(tab)*100)['2'] %>% nn(1)` percent of journals. The most labels any journal bears is `r names(tab) %>% tail(1) %>% as.integer %>% nn`, but these are outliers with most journals bearing only a few labels.

```{r mod-proj,include=T,cache=T,dev='png',fig.width=if(docx) 6 else 3,fig.height=3,out.width='30%',fig.show = 'hold',fig.cap='Mode Conversion on a 300 Edge Random Sample of the JSTOR Title List Label Network'}
withr::with_seed(12345,bup<-jclu2bup.f(jclu))
```

It is worth noting a few features of the unimodal projections or facets illustrated in the second and third panels. First, unimodal projections will always be made of overlapping cliques. Take the journal facet; each journal bearing a particular label will be tied to each other journal with the same label. Together they will form a clique, a subnetwork of maximum density where all possible ties exist. Such cliques grow nearly exponentially, as each additional journal with the same label joins the clique and adds a number of ties equal to the former size of the clique. In practice this means that very common labels like "Social Sciences" can easily dominate the unimodal projection of the network. Here the weighting of edges becomes important; if two labels overlap because some nodes bear both labels, then within the intersection of the two cliques the ties may be treated as "weighing more" by adding the contribution of each label separately. The exception is if the cliques overlap by only one node, in which case they have a node but no ties in common. Nevertheless using methods that take edge weights into account is a good way to ameliorate the exponential influence of popular labels.

Second, though the unimodal facets of a bimodal network represent the same data, each may have different characteristics especially in the common case of a large population imbalance between modes. In the full network there are `r jclu$b %>% {V(.)$type} %>% {sum(!.)/sum(.)} %>% nn` times as many journals as labels and each journal sends multiple ties. This degree imbalance between the two modes may mean that one facet is more dense than its inverse. Density is the proportion of actual ties out of all possible ties. In Figure \@ref(fig:mod-proj) an imbalance may be observed where the journal facet has many dense free floating or overlapping cliques and where the label network appears to be mostly made of isolated labels save for the few larger components. In the sampled network the journal facet is `r lapply(bup[ec('proj1,proj2')],igraph::graph.density) %>% do.call(what='/') %>% nn` times more dense than the label facet. In the case of our full network, the potential imbalance in degree distribution between facets happens to be offset by the population imbalance itself. The densities in the full journal and label facets are comparable, `r (sapply(FUN=igraph::graph.density,jclu[ec('gt,gd')])*100) %>% nn(1) %>% nl` percent respectively, meaning that analysis will not merely hinge on which facet is analyzed.

Third, unimodal projection has the effect of pruning what are sometimes referred to as pendants, which are simply nodes with only a single tie. Each of the isolates in the label facet represents a larger or smaller number of journals, which may be observed in the different sizes of the free floating cliques of the journal facet, yet no matter their size they supply no information about interdisciplinarity. Because the journal facet captures both size (of cliques) and relatedness (clique overlap) it is a better representation of the information of the original bimodal network. Its drawback is that it is larger and more unwieldy to analyze. The label facet offers a simpler picture of interdisciplinarity.

## Network Community Detection {#ncd}

Each facet described above will help answer a different question about disciplinarity in the JSTOR archive as indicated by JSTOR's labeling policy. I aim to resolve the uncertainty about which labels count as superdisciplines and to reveal patterns of sorting not apparent in the labels themselves. The rationale for doing this is to observe not the choices of JSTOR coders, but the tacit judgement they likely used in applying labels. I expect that the `r jclu$n$l` fine grained labels belie a simpler classification scheme of academic superdisciplines.

I will use two techniques, community detection and graph visualization, to answer these questions. Communities are really subnetworks of high density, or clusters. I operationalize disciplinarity as the presence of clusters within the journal facet network. Community detection on the journal facet will answer how many superdisciplines there are and the size of each in terms of the number of journals belonging to it. Visualization of the label facet will show how hard or soft are the boundaries between disciplines and where the strongest interdisciplinary relationships lay.

First, I use community detection to partition the JSTOR journals into mutually exclusive disciplines. Community detection is a set of network methods designed to expose clusters by grouping nodes together such that they send more ties to members of their own group than they send to members of different groups. There is a cottage industry around developing algorithms and statistical models to learn an unobserved community structure of a network [see @Fortunato2016Community for an excellent review]. The choice of the right community detection method is controversial especially for very large networks in which cross-validation is difficult. Fortunately the network at hand is small enough to validate directly which lowers the risks of choosing the wrong method.

To wit I adopt the well-known Louvain method of community detection based on hierarchical modularity maximization. [@Blondel2008Fast] Modularity is a quality metric quantifying the trade off between within-group and between-group ties. The modularity of any given partition of a network into clusters is equal to the proportion of ties that fall within clusters minus the expected proportion of within-group ties if ties were distributed randomly. A division that is as good as chance would have a modularity value of zero, a division better than chance a value between zero and one, and a division worse than chance a value between negative one and zero [@Newman2004Finding\:8]. Higher modularity scores indicate a better sorting of the network into densely tied clusters.

The Louvain method is a bottom-up agglomerative algorithm. The procedure starts by assigning each node to its own community. Then, for each node, it assigns the node to the neighbor's group that would most improve global modularity. It repeats this until no move improves modularity. This forms the first layer in the hierarchy. It then collapses groups into nodes and repeats the algorithm on the condensed network, stopping at the first level where there is no modularity improving move to make. The first layer represents the most local, the last layer the most global resolution of community structure.

Modularity-based methods are tried and true, and their drawbacks are well-known. The Louvain method is not deterministic, as the outcome may (but usually does not) depend on the ordering of the nodes in the reassignment queue. However Louvain has several features that recommend it. It is computationally fast on small to medium graphs and it is freely available in network analysis software. It also gives a hierarchical solution that provides the analyst with options to inspect community structure at a range of local and global resolutions, akin to a cartography of counties versus one of continents. Given the small size of our network, a local resolution will not be overwhelming, so Louvain is preferable to other methods that only offer the coarser global view.

Table \@ref(tab:jclu-tab-sup) summarizes the results of applying the Louvain method to the journal facet and taking the most localized layer of the community structure. Learned labels are applied to the clusters by assigning each the name of its most frequent label. Community detection sharpens the boundaries between fields by placing each journal unambiguously in one superdiscipline or another. This mutual exclusivity is apparent by the sum of the given labels exceeding 100 percent.

```{r jclu-tab-sup,include=T}
sg(jclu$tab %>% copy %>% setnames(ec('super,N,Labeled,LPct'),ec('Superdiscipline,Learned,Given,GPct'))
   ,tit = "JSTOR Journal Counts")
```

The first finding is that of the `r nn(jclu$n$l)` labels these `r nn(jclu$n$s)` form the top of a hierarchy of superdisciplines. Area Studies stands apart and is not subsumed under either Social Sciences or Humanities. Social Sciences journals predominate due to JSTOR's initial focus in that area, even without counting economics among them, and Science & Mathematics counts for a larger share than one might think. Economics stands apart from the Social Sciences, and indeed Business & Economics marks the transition from the larger academic journal space to the smaller professional space of Arts, Law, and Medicine & Allied Health.

The given labels do overlap and one can recover a picture of interdisciplinarity by clustering and visualizing the label facet. This facet presents a simplified view. Recall that each facet represents the same data, the difference being whether a journal or a label is represented as a node or an edge, and that there is a population imbalance in favor of journals over labels. The larger the population the easier it is to partition into a greater number of subpopulations. Conversely, because there are far fewer labels than journals, one would expect the clustering to be less granular for the label network than for the journal network. In fact there `r wrk<-lapply(jclu[ec('title_history,discipline')],function(x) {x$memberships[1,]} %>% unique %>% length) %>% do.call(what='-'); if(wrk==1) 'is' else 'are'` only `r wrk %>% nn` `r if(wrk==1) 'less' else ec('fewer,more')[(wrk<0)+1]` cluster`r if(abs(wrk)>1) 's'`--`r do.call(setdiff,{wrk1<-lapply(jclu[ec('ft,fd')] %>% unname,levels);if(wrk<0) rev(wrk1) else wrk1})`--which is subsumed under Social Sciences.

## Network Visualization

Figure \@ref(fig:jclu-lnet) visualizes the relationships among disciplines, where again the strength of ties is equal to the number of journals bearing both labels. Here the label with highest number of ties within its cluster becomes the category name of the cluster. That label is then omitted as a node and is instead visualized as a color coding of its cluster, reflecting the special status of the superdiscipline labels.

Unlike traditional graph visualizations that are designed to be pleasing to the eye, this  one is drawn according to a statistical model called a latent position or latent space  model. It starts with a simple idea that the weight of the edges (the number of journals carrying both labels) is a count that follows a Poisson distribution. This distribution may be modeled by log-linear regression where the logarithm of the mean of the distribution is a linear function of an intercept term and covariates. What is interesting about the model is that the covariate of interest is treated as the distance between the nodes in an unobserved or latent space. The distance is treated as negative such that as nodes get closer together (as the negative distance increases) the count of the edge weight between them increases (technically the logarithm of the mean of the count increases).

It is an elegant idea, but estimating the model is complicated. The distances are metaphorical, and to realize them requires positing a euclidean space in which each node has coordinates. From the coordinates the distances can be easily calculated, but knowing which are the right coordinates requires a complicated estimation routine based on optimizing goodness of fit between guesses of the coordinates and the actual count data. The estimator begins with coordinates taken from the conventional Fruchterman Reingold layout algorithm and uses Markov Chain Monte Carlo simulation to converge toward the positions that optimally fit the latent space assumption [see @Krivitsky2008Fitting for details of the model, estimation, and software]. Even if the estimator does not arrive at a perfect solution it improves upon a conventional layout in the direction of meaningful, and not just pretty, aesthetics thereby helping the viewer to avoid artifacts and perceive real information about the network.

Another great feature of the latent space model is that it allows additional terms to be fit alongside the latent distances. It is possible to control for or net out the effect of nuisance terms like any other regression. As discussed above there is a concern about the undue effect of popular labels. I have already tried to remove the superdiscipline labels from the label network, preferring to represent them as color coded categories rather than nodes. Popular labels may still remain, however, and due to the exponential growth of ties during downmode conversion even a handful of them will have a disproportionate influence on the global layout of the graph.

This degree distortion can be controlled for by what is called a sociality term, which can be thought of as a measure of a node's popularity. A sociality term is a score for every node that if positive means a node is more attractive and if negative means a node is actually repulsive of ties. When viewing the positions of a latent space model also fit with a sociality term, the space will measure relatedness without the effects of popularity.

Figure \@ref(fig:jclu-lnet) plots the results of a latent space model on the label facet omitting superdiscipline nodes. Figure \@ref(fig:layouts) plots the same in an animation comparing the latent space and traditional Fruchterman Reingold layouts. I will interpret the substance of the latent space layout below, but first I will comment briefly on some of the differences between it and the traditional layout.

```{r jclu-net,eval=T,include=F,fig.cap='Discipline Network. Within-cluster ties are darker and between-cluster ties lighter.',fig.cap='Discipline Network.'}
f<-'d/b/jclu2net.RData'
if(file.exists(f)){
  load(f)
} else {
  jnet<-jclu2net.f(
    jclu
    ,layout='fruchtermanreingold'
    ,repulse.rad=log(gorder(jclu$gd))*gorder(jclu$gd)^2 # reign in isolates, smaller is more compact
    ,niter=1e5
    ,isolates = F
  )
  save(jnet,file=f)
}
rm(f)
#plt(jnet,tooltip='text') # TODO fix tooltips to omit edge tips
```

```{r jclu-lnet,include=T,fig.cap='Discipline Network in Latent Space.'}
#fig.cap='Discipline Network in Latent Space. Node size represents sociality. The larger a node, the more attractive it is, and the larger a white dot within a node, the more repulsive it is.',
f<-'d/b/jclu2lnet.RData'
if(file.exists(f)){
  load(f)
} else {
  jlnet<-jclu2net.f(
    jclu
    ,layout='latent'
    ,isolates = F
    # ,burnin=1e4
    # ,sample.size=1e3
    # ,interval=1e2
    ,burnin=1e7
    ,sample.size=1e3
    ,interval=1e4
  )
  save(jlnet,file='d/b/jclu2lnet.RData')
}
rm(f)
if(latex|docx) {jlnet + scale_x_continuous(trans='reverse') + scale_y_continuous(trans='reverse')} else plt(jlnet,hovinf = 'text',tooltip='text') %>% plotly::layout(xaxis = list(autorange = "reversed")) %>% plotly::layout(yaxis = list(autorange = "reversed"))
```

```{r lnetg,eval=F}
f<-'d/p/lnetg.RData'
if(file.exists(f)){
  load(f)
} else {
  p=4
  t=1
  i=1e4
  s=max(600*t,1000)
  b=250*i*p
  cat('i =',i,'b =',b,'\n')
  g<-(attributes(jnet)$zkref %>% unique %>% length):1
  names(g)<-g
  lnetg<-pbapply::pblapply(g, function(x)
    try(jclu2net.f(
      jclu
      ,layout='latent'
      ,isolates = F
      ,grp = x
      ,burnin=b
      ,interval=i
      ,pilot.runs=p
      ,sample.size=s
      ,store.burnin=T
      ,threads=t
    ))
    ,cl = 7
  )
  save(obj,file=f)
}
rm(f)
pdf('_book/lnetg-diag.pdf')
lapply(lnetg,function(x) latentnet::bic.ergmm(attr(x,'lnet'))$overall) %>% {plot(x=names(.) %>% as.integer,y=.,type='b')}
withr::with_package('latentnet',lapply(lnetg[!sapply(lnetg,is.null)],function(x) attr(x,'lnet') %>% mcmc.diagnostics))
dev.off()
```

```{r lnet-frames,eval=F}
lnetg2pfp.f<-function(lnetg){
lnetg<-lnetg[order(names(lnetg) %>% as.numeric)]
prc<-list()
pdf('_book/lnetg-plt.pdf')
prc[[1]]<-data.table(
  1
  ,attr(lnetg[[1]],'net') %v% 'vertex.names'
  ,attr(lnetg[[1]],'net') %v% 'memb'
  ,attr(lnetg[[1]],'lnet')$mkl$sociality %>% exp
  ,attr(lnetg[[1]],'lnet')$mkl$Z.K
  ,attr(lnetg[[1]],'lnet')$mkl$mbc$Z.pZK %>% apply(1,max)
  ,latentnet:::plot.ergmm(attr(lnetg[[1]],'lnet'),Z.ref = attr(jnet,'zref'))
) %>% setnames(ec('f,vertex.names,memb,rs,zk,part,x,y')) # part = partisanship
for(i in 2:length(lnetg)) prc[[i]]<-data.table(
  i
  ,attr(lnetg[[i]],'net') %v% 'vertex.names'
  ,attr(lnetg[[i]],'net') %v% 'memb'
  ,attr(lnetg[[i]],'lnet')$mkl$sociality %>% exp
  ,attr(lnetg[[i]],'lnet')$mkl$Z.K
  ,attr(lnetg[[i]],'lnet')$mkl$mbc$Z.pZK %>% apply(1,max)
  ,latentnet:::plot.ergmm(x = attr(lnetg[[i]],'lnet'),Z.ref = prc[[i-1]][,cbind(x,y)],Z.K.ref = prc[[i-1]][,zk])
) %>% setnames(ec('f,vertex.names,memb,rs,zk,part,x,y'))
dev.off()
prc<-rbindlist(prc)
prc<-rbindlist(list(
  data.table(
    f=0
    ,vertex.names=attr(jnet,'net')%v%'vertex.names'
    ,memb=attr(jnet,'net')%v%'memb'
    ,rs=attr(jnet,'net')%v%'rs'
    ,zk=attr(jnet,'net')%v%'memb' %>% factor(levels=jclu$fd %>% levels) %>% as.numeric
    ,part=1
    ,attr(jnet,'zref')) %>% setnames(ec('V1,V2'),ec('x,y'))
,prc))[,zk:=factor(zk,levels=table(zk) %>% sort(T) %>% names %>% as.integer) %>% as.double,by=f][,zk:=factor(zk)]
scl<-prc[f>0,range(rs)]
prc[f==0,rs:=lintran(rs,range(rs),scl)]
prc[,x:=scale(x),by=f]
prc[,y:=scale(y),by=f]
prc[f!=0,x:=prc$x[prc$f==1]]
prc[f!=0,y:=prc$y[prc$f==1]]
setkey(prc,f)
pfp<-prc[.(c(0,1))][,f:=ec('Fruchterman\nReingold,Latent\nSpace')[f+1]][,memb:=factor(memb,levels=levels(jclu$fd))]
fp<-myth(ggplot(pfp,aes(x=x,y=y
#                        ,alpha=part
                        ,color=memb,ids=vertex.names
,size=rs,text=vertex.names,frame=f)) + geom_point(aes(stroke=1+rs/max(rs)*5),size=2,shape=21,fill='white',position = position_jitter(width=.01,height=0,seed=12345)
) + guides(alpha=F,size=F)) + theme_blank()
plt(fp,tooltip='text') %>% plotly::animation_opts(redraw=T,frame = 2* 1e3,transition=2* 1e3)
}
```

```{r testplot,eval=F}


net<-attr(lnetg[[1]],'net')
e<-net %>% as.edgelist
vn<-attributes(e)$vnames
e<-data.table(e) %>% setnames(ec('s,r'))
e[,`:=`(s=vn[s],r=vn[r])]
e[,weight:=net%e%'weight']
e[,cross:=net%e%'cross']
ct<-RColorBrewer::brewer.pal(unique(pfp$memb) %>% length,'Dark2')
names(ct)<-table(pfp$memb) %>% sort(T) %>% names

setkey(pfp,vertex.names)
pfp[,clr:=ct[memb]]
e[,`:=`(
  x=pfp[s,x]
  ,xend=pfp[r,x]
  ,y=pfp[s,y]
  ,yend=pfp[r,y]
  ,xm=pfp[s,memb]
  ,ym=pfp[r,memb]
)]
e[,clr:=mapply(function(s,r) colorRampPalette(c(s,r))(3)[2],s=ct[xm],r=ct[ym])]
}

cp<-ggplot() +
  geom_segment(data=e,mapping=aes(x=x,y=y,xend=xend,yend=yend,group=clr),alpha=.1,color=e$clr,size=e$weight/max(e$weight)) +
    geom_point(data=v,mapping=aes(x=x,y=y,text=name,size=rs),shape=21,color=desat(v$clr,.01),fill=v$clr)
cp

```

(ref:layouts) Fruchterman Reingold and latent space layouts compared. [*Interactive pop-out.*](`r if(latex|docx) 'https://brooksambrose.github.io/portfolio/'`exh/layouts.html){target="_blank"}

```{r layouts,include=T,fig.cap='(ref:layouts)'}
f<-'d/b/pp.RData'
if(file.exists(f)){
  load(f)
} else {
  load('d/b/jclu2net.RData')
  load('d/p/lnetg.RData')
  fr<-data.table(attributes(jnet)$ggnet)[,f:='FR']
  ls<-data.table(attributes(lnetg$`1`)$ggnet)[,f:='LS']
  x<-rbindlist(list(fr,ls))[,n:=is.na(weight)]
  x[(n),sym:=(rs>=1)+1 #%>% {ec('diamond,circle')[.]}
    ]
  
  withr::with_package('plotly',pp<-plot_ly(x=~x,y=~y,frame=~f) %>%
    #add_segments(data=x[(!n)],xend=~xend,yend=~yend) %>%
    add_markers(data=x[(n)],color=~memb,ids=~vertex.names,text=~vertex.names,size=~rs,symbol=~sym))
  save(pp,file=f)
}
rm(f)
class(pp$x$highlight$off) <-'character'
htmlwidgets::saveWidget(pp,file = getwd() %>% paste('exh','layouts.html',sep=.Platform$file.sep),background='white',selfcontained = F)
# why can't I fix the flipped axes!
if(latex|docx) knitr::include_graphics('img/layouts.png') else pp # %>% plotly::layout(xaxis = list(range = "reversed"),yaxis = list(autorange = "reversed"))
```

In the animation nodes are hollow shapes to make their overlaps more obvious. The size and shape of the nodes relates to their fitted sociality: circles indicate attractive and diamonds repulsive nodes, and the the larger the shape the more attractive and the smaller the more repulsive. The layouts differ in two important respects. First the global patterns are dissimilar. The latent graph has four discernible arms. The traditional graph is a cloud that at first glance seems evenly dispersed, however there is a prominent core periphery pattern, with circles concentrated in the center and diamonds displaced in a ring around them. Second, patterns of node proximity are distinct. In the latent graph, nodes may share or almost share a center, indicating structural equivalence. In the traditional layout concentric node positions are algorithmically prohibited by a repulsion radius, with apparent overlaps due only to the scaling of node size. The traditional core periphery pattern is the consequence of the imbalance between low and high degree nodes; by netting it out with the sociality term the latent model makes it possible for minor disciplines to cluster around the major ones to which they are anchored.

Table \@ref(tab:jclum) reports the disciplines by their superdiscipline as well as their sociality, which is reported as the exponentiated regression coefficient. The sociality term is a multiplier, which means that to predict the number of journals bearing the label of each of two disciplines you multiply the average or expected number of ties by the sociality term of the first discipline and again by the term of the second discipline. When the multiplier is greater than zero we expect more ties, when less than zero fewer ties. Does this mean we should expect the most ties to occur between the nodes with the two largest sociality terms, here business and political science? Not necessarily, because these effects are also combined with a term that is the inverse of the probability of a tie independent of sociality, which is given by the distance between their positions in the latent space. Because business and political science are not very close together, the big effect of their combined sociality is offset by the longer distance between them.

```{r jclum,include=T}
jclum<-withr::with_package('network',attributes(jlnet)$net %>% {data.table(m=.%v%'memb',n=network.vertex.names(.),rs=.%v%'rs')}) %>% setorder(m,-rs)
jclum[,rs:=round(rs,2)]
jclum<-jclum[!'Medicine & Allied Health',on='m']
jclum['Management & Organizational Behavior',on='n',n:='Management & Org. Behavior']
jclum['Architecture & Architectural History',on='n',n:='Architecture & Arch. History']

jclum<-lapply(
  unique(jclum$m),function(v) {
    M<-jclum[m==v&rs>1,.(n,rs)]
    if(!nrow(M)) M<-data.table(NA_character_,NA)
    m<-jclum[m==v&rs<=1,.(n,rs)]
    if(!nrow(m)) m<-data.table(NA_character_,NA)
    v<-data.table(M,m)[
      ,lapply(.SD,function(x) replace(x,duplicated(x),NA))] %>%
      {data.table(Super=v,.)[,Super:=replace(Super,duplicated(Super),'')]}
    rbindlist(list(v,rep(NA,ncol(v)) %>% as.list))
  }
) %>% rbindlist %>% setnames(ec('Superdiscipline,Major Discipline,Soc.,Minor Discipline,Soc. ')) %>% .[!.N]

sg(jclum,tit='Disciplines by sociality multiplier',font.size='small'
   #,note='Disciplines with sociality < 1 have fewer > 1 have more ties than expected by chance.'
   )

#jclum[,m:=replace(m,which(duplicated(m)),'')]
# jclum<-jclum[,.('Discipline (Sociality)'=paste0(n,' (',rs,')',collapse=', ')),by=m] %>% setnames('m','Super')
# withr::with_package('xtable',print(xtable(
  # jclum
  # ,caption='Disciplines'
  # ,label=knitr::opts_current$get('label')
  # ,align='llX'
  # )
  # ,tabular.environment = "tabularx"
  # ,width="\\textwidth"))
```

## Discussion

Recall that the superdiscipline classifications are functions of their underlying subgraph densities, hence we should expect them to be clustered together in a layout that represents relatedness. For convenience we will refer to the cardinal directions to locate ourselves on the plot. The most prominent feature of the global patterns is the appearance of two axes. From west to east is an axis dividing Science & Mathematics and the Social Sciences. From north to south is a more diffuse axis dividing Business & Economics from the Arts and Humanities. Though it is tempting to interpret the south to north axis as artistic opposition to profit, it is more plausible that they are each facing a different side of the Social Sciences, which neatly bifurcates them.
While superdiscipline members are usually near each other, they do not all bunch together in the same way. In the north Business & Economics is the most spherical cluster with its members much more strongly connected inside than outside the group. By comparison Science & Mathematics is extended between two internal poles, with life sciences pulled far to the west and statistics and mathematics interfacing with other central disciplines. The Social Sciences are similarly extended but with three poles: politics in the east, archaeology in the south, and geography in the west. Area Studies, which subsumes history, tends to closely overlap the Humanities with the exceptions of asian and middle east studies to the east, which are situated between political science and archaeology, American studies in the west, and the history of science and technology, which is closer to its subject matter in the west. Most of the Humanities is highly concentrated with the exceptions of bibliometry and library science on one hand and classical studies, which is close to archaeology, on the other. The arts is the most diffuse superdiscipline; several of its fields are on the margins with the exceptions of urban studies at the confluence of Science & Mathematics and the Social Sciences and music as well as performing arts sharing the interdisciplinary center with the Humanities and Area Studies. Finally, the three disciplines in Medicine & Allied Health are split between a Science (public health, health sciences) and Social Science (health policy) orientation.

There are several interesting cases of total or near structural equivalence. Total overlap for major fields suggests extra superdiscipline candidates, with biological sciences covering zoology and history covering American studies. Considering only minor discipline overlaps, the public health and health sciences labels overlap to the point of redundancy, as do library science and bibliography. Near overlaps between minor and major fields may indicate a theoretical (major) applied (minor) relationship. This is expectation is confirmed by the major law and minor criminology & criminal justice overlap, and contradicted by the major education and minor psychology overlap, where the minor psychology field would be considered the more theoretical. Interestingly, structural equivalence of fields in two different superdisciplines occurs only once, with religion (Humanities) and Jewish studies (Area Studies) tightly coupled.

A last topological feature of interest may strain the accuracy of the model, but I lodge it as a problem for further consideration. Within Area Studies there appears to be two differentiated fronts, a more southerly line of "white studies"--history, American studies, British studies, Irish studies, and Slavic studies--and a more northerly line of "brown studies"--American Indian studies, African American studies, Latin American studies, and African studies, which also includes a spot for feminist and women's studies from the Social Sciences. The distances in question are very close, and model variability may reveal them to be statistically insignificant. Even if they are significant their overinterpretation risks a narcissism of small differences.

```{r pwn}
f<-'d/b/pwn-dis.RData'
if(file.exists(f)){
  load(f)
} else {
  pwn<-network::as.edgelist(attributes(jlnet)$net,'weight') %>% {data.table(.)[,`:=`(V1=attr(.,'vnames')[V1],V2=attr(.,'vnames')[V2])]} %>% setnames(ec('s,r,ew')) %>% setkey(s,r)
pwn[,th:=mapply(function(x,y) c(x,y) %>% {jstorm[!duplicated(title_history)]$title_history[sapply(jstorm[!duplicated(title_history)]$discipline,function(x) all(.%in%x))]},x=s,y=r)]

  save(pwn,file=f)
}
rm(f)
```

## Cartography of Classification

There are many stories to tell about a map that we trust to be an accurate metaphor for closeness or relatedness in a cultural field. Recall that statistically, the closer two nodes are the greater the number of journals listing each as a subject category. The decision of how the subject labels were applied was a policy decision at JSTOR using an unknown procedure. Our interpretation of the map is affected by how credible we think the archivists' decisions were. Without knowing about their process, it becomes a matter of trust. What the present study does is provide a framework in which the cognitive act of labeling can be represented. Representation in turn opens up opportunities to compare multiple representations, and to ask questions about how competing systems may influence each other.

The real value of the methods outlined above is that it helps to represent the global patterns of local decisions. Researchers often treat genre classifications as given features of language and the culture of their time. The assumption here is that there is a classification system held in common by all actors in a culture. Yet this does not necessarily comport with claims that genre systems are also fluid and contested. Are they contested out of a deliberate attempt to subvert the status quo, or are they contested out of error or ignorance of the same? Do disagreements simply reflect preferences for an alternative classification?

A question for further research is whether the archivist's cognitive map is the same one that one of their role partners, for instance a scholar or a student, would share. We might expect that because this map given by JSTOR is at a medium level of granularity--compared for instance to the thousands of possible subclassifications within each superdiscipline, or the even coarser distinction between art and science--that it may have a big influence on neophytes who may have little experience with scholarship, but that it may be resisted by experts who have a more complicated view of their own domain.  The ability to bridge these very different levels of expertise may itself explain how archivists would arrive at this particular resolution of categories.

The next phase of this research then requires the acquisition of more labeling systems, and the generation of more maps. This will no doubt open up new methodological problems and opportunities, such as whether the morphology of a particular map can be matched to another even if the underlying labels are different. I hope that this first step opens the door to new problems whose study will bring us closer to understanding what genre systems really are and the effect they have on our understanding and use of knowledge.
