==> 11.Rmd <==
<!-- `r tpe<-'Profession against Education: Knowledge as Influence'` -->

<!-- #### Chapter 1: `r tpe` {-} -->

<!-- (ref:pe-abs) Status reflects a person's generalized ability or inability to accrue resources to herself over the claims of others. Social institutions define the conditions of status attainment, and different institutions validate statuses according to different norms, truths, and values. Knowledge as a valued resource forms a basis of status hierarchies in two institutions in American society: profession and education. This essay explains how knowledge mediates social relationships to stabilize status hierarchies differently in professions and in education. It also describes how stability of status hierarchies depends on a qualified segregation of these institutions, as the norms of one can be corrosive to the norms of the other. -->

<!-- (ref:pe-abs) -->

# `r {load('_init.RData');kopi();tpe}` {#pe}

#### Abstract {-}

(ref:pe-abs)

#### Keywords {-}

yaba, daba, do

## Pitch

### Etymology

Education is from the Latin "to lead out", while profession is "to declare publicly". Knowledge establishes the parameters by which a teacher educes a student; the only way a student may be intentionally developed is to understand one's own knowledge as a standard against which a student's may be evaluated. There is a delta between the two and one may be brought nearer the other by a series of transformations.

Knowledge transforms in three simple ways: addition, deletion, and replacement.  The teacher's knowledge acts as a template or paradigm by which the student's may be compared. Under unregulated relations teachers may adapt their own knowledge in interactions with students such that the two knowledges converge at a new level albeit closer to the teacher's starting point. In schools however the institutional force resists adaptation, so knowledge transformations tend to be anchored forcefully to the formal program. Rather than representing the teacher's pole, the program is a third pole effecting both teacher and student. This regulation both conserves and ossifies the knowledge, which promotes consistency across time and personnel but which risks alienating supporters with current projects.

### Antecedents

Learning theory, cognitivism, behaviorism, constructivism. We would be communicationist.

## Theory

### Statuses

To introduce the concept of status it will be useful to consider the most important case, which is itself constituitive of U.S. society.

American society is constituted by five ascriptive statuses: age, gender, race, class, and language group. These statuses depend on trivial and near instant recognition in mediated social interactions, hence their grounding in perceivable physical characteristics. The ease with which new members of society are trained in the perception and classification of these statuses lends them stability. The institution they form is American society writ large as no person participating "in America" can escape their bearing. Age, gender, and race are usually discernible by simply witnessing another person's body. While class and language group sometimes not visually confirmable, they are just as trivially known the moment they become salient, when the use of language or mobilization of wealth is required for social interaction.

That some American subsocieties attempt to abrograte these norms demonstrates rather than contradicts their saliency. Constitutional statuses, while never natural or inborn, are nonetheless historical, changing only on a societal or generational scale. So long as the norms are easier to learn than to forget they are reinforced.

The classification of people by these constitutive statuses is a "cold" cognitive operation concerning definitions of reality. These statuses are social facts deeply embedded in the consciousnesses of experienced members of American society. The ordering of these statuses is a more complicated "warm" mixture of cognition and emotion concerning not only definitons of reality but affective values as well. ; it is a rigid, though less so than the classifications themselves.

Their ubiquity in American society renders them foundational for other institutions more limited in scope, thus all other institutions are shot through with the effects of these societal statuses.

### How Knowledge Works

Knowledge in general is both a content and a capacity. As particular content it is the result of human learning and forgetting that varies historically.  As a general capacity it is the mechanisms of learning and forgetting themselves that through constant operation yield the content of human experience. Knowledge is  a capacity with which all people are endowed and which operates naturally as a mechanism of human accomplishment.

Human knowledge is sometimes nothing but the cognition of other animals who encounter and adapt to the world according to their own capacities for learning. It is usually, however, something more. Human knowledge can be codified outside of the mind, and such codifications allow transfers between people and across time.  Codifications presuppose a code, the most primitive of which are human languages. Knowledge is complex in that smaller components hang together in larger groups. Meaning relates components and regulates what may and may not belong to a larger group. Meaning renders knowledge useful by linking problems to solutions. Culture is the social aspect of meaning; culture happens when people regulate each other's expression as meaningful or not.

#### Ideas

Call an idea a component of a knowledge. Call a knowledge a particular configuration of ideas. Both are psychological units that cannot exist outside of a mind. Knowledge is composable in that its smaller ideas may be inspected, added or deleted from configurations, or reconfigured in their psychological relations. Learning is the time and effort required to understand an idea or to reconfigure a knowledge.

Call meaning the morphology of ideas of a particular knowledge. Ideas relate in a pattern of dependencies, where the rank of an idea is the sum of recursive dependencies in its longest chain of ancestors. Depth is the number of ideas constituting a knowledge, weighted by the degree of their dependencies.

#### Language, Codification, and Encryption

Call a text a symbolic expression of a knowledge, and a token a symbolic expression of an idea. A code is a transformation program allowing for the reliable inscription of an idea into a token or description of a token into an idea. A language is a configuration of codes.

knowledge:language::idea:code

A cipher is a token that has been recoded at least once, usually to segregate cultures within the same language. A token and a cipher express the same idea in different symbols.

Encryption is illegible codification, usually done to render a token inert in transmission. It differs from ciphering in that it cannot be understood directly without decryption. Encryptions are inert in that they do not express the idea, whereas ciphers do.

#### Communication and Culture

Call a communication the simultaneous expression of an idea in the mind and a token in language. A communication is the couplet of the two. Communication allows tokens to act as hinges between minds, if a token evokes the same idea in each.

Call a population of knowers a culture. Cultures embed knowledge with the arbitrary and conservative mechanisms that guarantee consistency across people and time. A quorum is a collection of the smallest number of knowers necessary to compel a marginal deviant to comply with the meaning of a knowledge. A paradigm a textual standard used by a quorum as a template against which a test text may be compared.

#### Corollaries

This can be summed up in a simple theorem: the speed of the growth of knowledge is inverse to the size of the culture that knows it. Ceteris paribus, knowledge that grows in content shrinks in understanding.

It does not necessarily follow that the shallowest cultures are also the largest. Here it is important to distinguish between being useful and being interesting. Knowledge is functional to human goals, so useless knowledge is easily ignored, and ignorance is the death knell of a culture. Shallow and pointless trivia will not support a culture. Knowledge that is useless may however still be interesting. In this sense the knowledge is directly consumed, it is the goal rather than a means to it.

Whether a component of knowledge is factor or a product depends on its meaning with reference to a person's project. As a project changes, so may the meaning of an idea. And interpretations change across individuals and within the same individual over time.

#### Knowledge Transactions

A culture is like a genotype in that it is a pattern that is repeated within a bimodal population. As Luhmann reminds us, minds cannot connect directly to each other, but must do so through a medium of communication. Hence knowledge exists always as a couplet between a mind and some sort of text. A mind without text is a culture of one, inscrutable to anyone including perhaps the knower herself. Thoughts not written down may be forgotten, or perhaps more to the point, durable and persistent thoughts will eventually be "written down". Inscription need not be formal; indeed it can be any empirical pattern outside of the mind that serves to replicate a pattern within the mind.

#### Structural Model

The functions of education and profession vary in ratios across organizations that may themselves be wholesale categorized as one or the other.

A professional practice is, in its client relations, almost entirely professional without education. Indeed a desire to learn on the client's part may be treated perfunctorily, as a marketing problem rather than a teaching problem. Convenient simplifications or misdirections will be offered to secure fees for services, as it would be a mockery to suggest that what the professional knows could be transferred transactionally.

#### Typology

Realistic locations on the ratio between educing and professing can be simplified into three types. A dominance of p is nothing but orthodoxy, while that of e is orthopraxy. A balance between them is rare and it is our task to understand it. The fourth option is the institutional neutralization of the knowledge dynamic altogether.

A professor uses knowledge asymmetries to reinforce status, and she will be careful to deliver a sterile version of messages that are useful only within the scope of a client transaction. Learning beyond the scope amounts to giving away the goods, but more threatening still would be the demystification of the professors status in the mind of the clients. There is thus a constitutional incentive to maintain the ignorance of the laity. This may but need not be done perniciously; indeed only a special type of client, a student, expects to receive knowledge of the quality emic to professional practice. Laity are busy pursuing the goals that led them to contract the professional services to begin with.

Codification of knowledge into different qualities is important here. The first split is between productive and unproductive knowledge, or consumed rather than invested knowledge. Learning as entertainment means messages are not intended to be used as means to an end, even if memory allows it to be repurposed later. The double contingency of intention means that sender and receiver may have different expectations about the economic meaning of messages.

Schools use institutions to resolve the uncertainty of cultural knowledge relationships. They replace double contingency with heteronomous expectations for both students and professors. Even if the intention around a message is shifting across or within supporters, schools create formal expectations that messages be treated as sanctioned work. This enables the institution to function predictably across personalities and across time. The arrangement works best for supporters vacillating between leisure and work modes, or who have no autonomous goals to contradict the institution, as may be the case where opportunities are scarce.

Contrary to appearance, schools cannot hitch their stability to knowledge itself, because interactions with it are very personal and unpredictable. Knowledge is sealed behind procedures, where assessment defines the messages that are relevant for rewards or punishments, which are only ever the tip of the iceberg, though savvy assessors may hope such facets are good indicators of what lies beneath. Hi

### Class System

It is a constant usually unobserved feature distinguishing professional organizations from corporations that knowledge is prevented from being sold as technology. This gives the prize away. It does not represent a lost economic opportunity only; the foundation of professional status may be threatened.

Professions are always more cryptic than is technologically possible. A profession that has been stripped, that has been rendered legible to laity, has lost its source of security, and prestige depends entirely on the esteem supporters lend to individuals. This is a tenuous situation.

Education concerns the development of knowledge within individuals, and in balance is a support to professions. Would an oversupply of education threaten the scarcity on which high fees or wages depend? Indeed it may not, if job growth grows slower than the educated class. Scarcity depends on the creation of jobs.

So social asymmetries are easily maintained where knowledge is on equal footing across supporters. This is clear, as much more goes into job acquisition than knowledge.

#### Class Terms

Let a discourse be a symbolically mediated conversation among more or less knowledgeable participants.

Call a knowledgeable participant in a discourse a supporter.

Call profession the maintenance of resource asymmetry among supporters into professors and laity.

Call education the maintenance of resource symmetry among supporters.

Call a student a supporter seeking to transition from laity to professor.

#### Authority

A measure of the quality of knowledge is available in the so called periodic table of knowledge.  First is "beer mat" knowledge, which is the lowest level. Here laity have no hope of contending with teachers or professors.  The knowledge relationship may be described as an asymmetry. With respect to a given item of knowledge, and this itself is problematic, the two parties to the relation stand in an authority conditioned by the distance or difference in the level of knowledge each commands.

#### Relationship of Professions to Knowledge

Professions are aided in their control over knowledge by natural excludability. Knowing is built on itself as a chain of dependencies which are inscrutable to laity. Professions excluded laity not normally by encrypting knowledge, though this is done, but by failing to make its dependencies transparent.

Education then is the structural inverse of profession. The task of education is to make knowledge scrutable. Education and profession can only be symbiotic in a careful balance, because an extreme of one will entail an erosion of the supports of the other. A great expansion of knowledge on the part of education may outpace job growth and turn a scarcity of skills into a glut, increasing competition and depressing wages in the profession. An encryption of knowledge or again the professional choice to not assist in the dissemination of know-how can choke off access to instructional resources, slowing the rate of knowledge transfer by educators.

Yet in a productive balance both education and profession can be mutually beneficial, each benefiting from a surplus of laity to support them in a mobility pattern. Education can install itself as the means of access to profession, reducing the costs of apprenticeship for professions. Resources for reproduction can be safely allocated to production.

#### self replication

While knowledge is the mechanism of professional exclusion, it may also be useful to think of the difference between e and p as an attitude toward self-replication. Professionals "subject" (cast down) clients on the dimension of knowledge while clients subject professionals on the dimension

#### Research against Teaching

Colleges and universities are organizations that exhibit a "basket weave" of both structures above. They use procedures to integrate the tendencies of each, such that a division of labor between profession and education may be stably maintained. By financing and facilitating both teachers and researchers, universities promote a productive balance between their competing tendencies. In creating forums where either profession or education happens, they also create stable opportunities for discursive antagonisms.

#### Discursive antagonism

Simply, professions and education experience each other as stressful because of inverse kinds of knowledge transactions.

Education is stressed by profession wherever professionals may render services in a way that black boxes the knowledge securing the service. Profession gives a person a fish, so she may live today and starve tomorrow. Education teaches a person to fish, so she may starve today and live tomorrow. As easily as they complement each other they replace each other. When both are present to laity, it becomes the consumers choice which path to invest in.

As the mobility pattern is the only way out of their conflict, the stage sequencing of education before profession is the tendency toward which universities develop.

### Functional Limits

A natural orientation toward knowledge is real driver of the existence of a profession/education structural dichotomy. There is a solution to the knowledge problem that would render the extant division of labor between education and profession obsolete. Natural excludability must be neutralized and replaced by a natural includability.

Cutting the knot means reducing the cost of bridging the gap between the presentation of a problem and the application of a solution. This could be done in two ways.

First, knowledge can be codified into technology such that human capital development is no longer necessary, i.e. the fool-proofing of all technology. Laity, and we would all be laity, would never need to learn in order to solve a task. This is the tack of all consumer-oriented product development.

This corporate nirvana engineers and distributors usurp the role of both professor and teacher. The technology replaces each, and no durable social relationship is required for the laity to solve their problem. A spot transaction to acquire the tech is all that is required.

Second, and on the opposite extreme, when confronting a problem laity could acquire the knowledge necessary to solve it immediately and costlessly. There would be natural includability. This is the tack of the internet search engine; one may express a problem in natural language and receive the program for mastering it immediately. There would be no social value to knowledge, professorial asymmetries among laity would be easily balanced, and thus no seeds for professions could be planted. Knowledge would be economically but not socially valuable. No professorial relationships could develop because all people would have equal status with regard to needed knowledge.

This is the communist nirvana envisioned by Karl Marx and the superhero scenario of NBC's The Pretender.

To argue that neither extreme is possible is to make a different assumption about the limits of human capacity for knowing. In scenario one, nobody could learn in interaction with technology, so we assume that consumers are hopelessly dim. In scenario two, consumers are unrealistically brilliant.

So the extremes are not real, but are historically unrealized ratios indeed unreal?

## Map

### Maps

There are two reasons to map knowledge spaces. First, we may want to know how knowledge develops as a resource unto itself. Second, we may want to exploit such a map for a productive purpose. Here we will attempt the second as prologue to the first. We will tackle the technical problems of constructing a map. We will show how a map can be put to use. Finally, we will investigate how the particular map we make may tend to predictably get us lost.

### Mechanism

What has been missing is a  mapping of exclusion mechanisms to the cultural content of knowledge. How do incumbent or pretenders use their mastery of knowledge to do boundary work?

## Story
### Remembering Social Science

Library technology, paradoxically, makes it easy to forget scholarship. [@Cevolini2016Forgetting] A scholar who is confident that knowledge is safely protected in archives and who believes that it can always be accurately retrieved is relieved of the urgency of digesting it now. 
Punting on understanding *from* texts opens space for exogenous understandings *of* texts. Contemporaneous cultures and societies constraining the reader prejudge the meanings of texts. While scholars genuflect to a quest for truth, the practical incentive to minimize time spent reading leads to several methods of collusion in the reduction of texts.

First, any text that can be commonly declared irrelevant need not be bothered with it all. Forgetting is the most efficient way to deal with a text.

Second, a text need not be read to be understood, as their meanings are common sense. Texts then have reputations and reputed meanings. These reputations like those of people stand in for the real thing in scholarly discourse. These mediated versions of the truth act as promissory notes in scholarly transactions; where the reputed meaning is accurately conveyed the issuing party is given credit for knowing the underlying truth, without burdening the transaction with a test of the truth. Like all symbolic currencies, the pace of commerce quickens where the symbol of the good is exchanged prior to the good itself needing to be mobilized.

The utility of reputed rather than real meanings functions to save scholars from the existential crisis besieging them: the pace of the publication of scholarship has long outpaced anyone's capacity to read it. The existence of subdisciplines is an obvious symptom of this trend bemoaned by every generation of scholars.

### Secondary selection

Primary selection is that of the editors and reviewers; of all work submitted for publication they have authority to choose what is made public. Secondary selection

*** 

***

> NB: *The remainder is unorganized but lays out general concepts relevant to cultural closure and social stratification*

***

## Ubiquity and Esotericism in Culture

Genre is a form of meta-expertise called *ubiquitous discrimination* [@Collins2007Rethinking\:45], which reflects a layperson's ability to make judgments (genre classifications) about what might otherwise be considered esoteric content. Ubiquitous knowledge is assumed to be available to all members of a society, even if not everyone necessarily acquires it. The availability of knowledge is a critical articulation point between cultural systems and social systems.

Bodies of knowledge and practice, cultures, exist because of yet apart from particular people. While some cultures are ubiquitous and inculcated in all members of the societies in which they are substantiated, others are esoteric and marked by significant barriers to learning them. Such esoteric cultures are as inaccessible as they are unnecessary. A child who does not learn what to eat and how to eat it in her local culture will have great difficulty doing anything else in her society. Food culture is of the ubiquitous type and is not difficult to learn. On the other hand, a child who does not master scripture will be ashamed at temple, but shame is bearable in a way that hunger is not. Scripture is of the esoteric type; knowledge of it enables special abilities and grants access to restricted aspects of a society, but ignorance of it does not threaten one's lay livelihood. Even in religiously totalitarian societies, what religious knowledge is necessary will be ubiquitous. What religious knowledge is esoteric can be safely left to the clergy. It is not too narrowly circular to say that the knowledge that informs daily life is the easiest to acquire, and that the mere act of living also reproduces that knowledge within contexts that are themselves ubiquitous. Whether relationships of ubiquity are familial, economic, or state relations varies by society, but no society lacks ubiquitous culture and the means of reproducing it.

Esoteric cultures, then, are removed from ubiquitous ones though they depend on them. They require specialized social relationships for their reproduction in populations. These special relationships may be called fields of cultural production (FCP). FCPs are both reproductive of extant knowledge and productive of novel knowledge relevant usually only to the field itself. To the extent that esoteric cultures are complicated or otherwise have high entropy, the failure to reproduce their cultural content is always an existential threat. While it is a guarantee that cultural content is constantly actually lost to reproduction failures, cultures that nonetheless persist historically must always have a socially structured FCP that, in inculcating new members and reinforcing the knowledge of extant members, resists the entropic decay that would otherwise lead to extinction.

The scope of a culture has a knowledge and a social dimension. Units of knowledge are both discrete and combinatoric. Discretely, the size of a culture can be measured by a count of co-occurring ideas and skills. The utility of knowledge, for instance as instantiated in an FCP, often depends on the combination and interaction of such discrete elements. The mass of knowledge in a culture is related to the count of units and to the ease with which those units can be learned, thought, or deployed in meaningful activity. Some cultures are easier than others, and it is the effort-adjusted size that is the real mass of the ideational content of a culture.

Not all sociologically relevant information is cultural. In the triadic relationship among culture, social structure, and personality, psychological concepts like schema that are sometimes treated as constitutive of social structure [@Sewell1992Theory] do not belong in the category of culture. Schemas allow personalities to organize their responses to a world that includes many things beside culture. In order for cultures and social structures to function stably in time, they depend on the successful operation of schemas, but they do not necessarily provision them. An only partial exception is the limiting and very specialized cases of pedagogical institutions. A culture can only exist if people can reproduce it, but the people must solve the psychological problems of cognition on their own. Whereas culture is fungible in its different mediated forms, there is no reason to believe that a "single schema" occurs across a population of people. There are no one-size-fits-all solutions to human motivation, and schemas will to a large extent be idiosyncratic, adapted as they must be to the particular circumstances in which people as träger find themselves.

What cultures can do to hedge their bets against the instability of personalities is to introduce kinds of knowledge adapted to cognitive problems. Schutz identifies systems of relevance from the side of personalities, and here we understand the same from the side of cultures.

People know culture and it does not appear apart from them. Put differently, culture is information and people are the media that concretely manifest it. The size of a culture is the count of the unique ideas and idea combinations appearing in a deduplicated media catalog, excluding possible but unrealized combinations. The potential of a culture is the size of the realized and unrealized combinations, where the number of elements in a set is limited by a given historical cognitive and technological memory capacity. The potential of a real culture is always greater than its size, as combination novelty always easily exceeds the available media.

The social mass of the culture is the enumeration of the media in which cultural content actually appears historically. The cultural mass is copied piecemeal among a population of knowers. The mass may be enlarged by information technology by making it easier to access and transmit culture, but the artifacts and technology themselves do not count in the mass of the culture. Real, historical human nervous systems, including language expression,

Dead cultures cannot live again because the dead part, the social part, is highly ephemeral. Even if a culture leaves material and symbolic artifacts, the "dead" parts are precisely those human relationships that would socialize new members in particular ways that cannot be deduced by people who never belonged to begin with, namely the archaeologists of the future. Such archaeologists belong to the FCP of their own time, and approach the historical artifacts of a dead culture as any other FCP approaches regular patterns in the world, as subject to their own particular cultural logics.

Reproduction does not necessarily imply long term consistency of cultural content, indeed a high level of knowledge loss may be a sign of historical stability in an esoteric culture if it results from a highly productive FCP. An oversupply of cultural content increases the chances that some, any culture can be conveyed into the future in a chain of descent.

Esoteric knowledge is essentially cryptic, ubiquitous knowledge essentially evident. Crypticism means that laity cannot acquire knowledge on their own, that is, without socially structured access to an FCP. Esotericism is the social consequence of crypticism. Cultures vary by knowledge features, as knowledge is a social axis of power along which status hierarchy necessarily develops. Knowledge asymmetry structures social relationships, usually leading to status inequality.

High entropy cultures must socialize new and maintain existing members like any culture, but their complications also tend to be generative of novelty.

The ubiquitous cultures will be learned because there is nothing in a society to do without knowing them, while esoteric ones are guaranteed to be forgotten by most, swapped easily for the always ready ubiquitous alternatives. Ubiquitous cultures have high ambient findability [@Morville2009Ambient], they can be transparently acquired for instance through mimicry [@DiMaggio1983Iron].

FCPs are social structures that can be thought of as the articulation points between people and a special type of culture that is associated with higher education.

To be sure, if esoteric cultures are marked by asymmetries of knowledge between laity and clergy, to continue with the religious example, one must admit that such asymmetries exist in ubiquitous culture as well, as in the supposed ignorance of a child against the learning of an adult. The ignorance of children combined with their megalomania is a useful example of how social structures operate. Social structures are always culturally conservative; a child's ignorance is a generative force, as they may attempt to solve problems in novel ways, and their efforts will frequently be corrected by the experts. Whether the correction takes marks the chagrin of the parent. Children quickly become specialists in their own FCPs, but this is never because they aren't also aware of the ubiquity. They may know it, it may bore them, and they efface it for fun. But they know what is ubiquitous; their proximity to ignorance of it may animate their creative abrogation of culture. The essentially conservative function of pedagogy will stamp out that creative flame in time. Indeed children are a great source of entropy for ubiquitous culture, hence the large investment in education generally.

This evolutionary argument can be more simply stated as, complicated cultures cannot exist for very long without stewardship.

## History

A final strategy for choosing what to read is a genealogical approach [@Foucault2002order]

# from 02

To return to disciplines, answering such questions would put the researcher in an adisciplinary predicament, for it would mean eschewing the scope restrictions cherished by disciples. The challenges of a meta-disciplinary analysis are manifold and uncertain. One is as likely to grow her audience as to lose it altogether. She risks the dilettantism of a jack-of-all-trades. What's worse, she exposes herself to a dizzying scale of content to consider. If disciplines are indeed functional, then the meta-analysis that effaces disciplines risks being dysfunctional. The upside, however, is appealing. If the universe of meanings given to the term genre does contain complementary uses, a meta-analysis will allow one to consider the consequences of the now arbitrary segregation of a superior metaconcept across disciplinary boundaries. Both sides of the divide could be strengthened by a cultural exchange of their respective terms. The redundancy of parallel discovery can be avoided.

In the sociology of culture, genre is a form of classification enacted by people in various social contexts. In the context of industrial capitalism, genre is an economic principle helping to organize supply and demand within markets for cultural products. There actors see genres among the borders between economic, social, and cultural uses, as a market category helping them acquire or produce content, as a card in proximate games of prestige, or as something to taste, to consume and enjoy directly. Less often genre is knowledge, a component of culture separate from taste, that is a factor in the formation of ideas and skills, whether these lead in turn to economic production or not.

Genre's meaning is context specific and variable across sociological subfields, though it is amplified in empiricist fashion by economy and society approaches that reduce genre to the act of classification itself. I say empiricist because of the empirical ease with which the classification or labeling actions are observed. Especially with Internet distribution systems, it becomes trivial for corporations to observe when a consumer labels her preferences while browsing a content catalog; it is much harder to observe the ideas that consuming a particular piece of content sparks in the mind.

In cultural studies genre is used much differently and much more in accordance with its etymology. To the humanist, genre is an ontological phenomenon, which is to say, genres are differentiated from each other by combinations of discrete features of signs and signifieds. Humanists are trained to establish these ontologies through methodological readings of texts and through cultivation of theories of genre types. These methods are at the same time empirical and interpretive, because in consuming objective texts the researcher actually observes the ideas that form in her own consciousness, and they may hope that others have a resonant experience. Genres have more substance to them for humanists than they do for sociologists. For sociologists of culture, genres are how people use genre labels, while for humanists, genres are knowledge.

As I have said, the meaning of genre for a sociologist of culture is economistic, at the same time a market category and a taste configuration for consumers. The term differs for a sociologist of knowledge, and perhaps for a cultural sociologist. It is ontology as it is for the humanist, however it is not the ontology as represented to the researcher in the consumption of texts. It is the hopelessly unobservable distribution of ontologies appearing in a population of consumers attending to similar texts. The sociologist of knowledge accuses the sociologist of culture of reducing internalist concepts, thought and experience, to externalist ones, taste and preference. The sociologist of culture rejoins, show me proof, and on and on the interlocutors spin around the axis defining the boundary between their subfields.

But this description of subdisciplinary differences really is just an example of a structural theory of genre that is within scope for the sociologist of knowledge and beyond it for the sociologist of culture, due to their epistemological differences. If genres are ontological, then they deeply structure a person's experience of reality. Ontologies form basic perceptual categories, and people with different ontologies of an object experience different things even if oriented to the same objective phenomenon. A ontological theory of genre would, for example, attempt to explain differences in taste as differences in phenomenological perception, whereas a taste theory of genre treats consumption behavior as revealing preferences whose downstream consequences are then explored.

Yet for all the differing treatments of genre mentioned above, do these views really contradict, or are they in fact complementary? Does genre as distribution serve the economic sociologist's goals, or does a lack of ontological substance lead her astray? Does genre as knowledge remain hopelessly unverifiable, or are theoretical constructs necessary to achieve a correct interpretation of facts? Is everyone at the club listening to the same song hearing the same thing?
