==> 12.Rmd <==
# Four Fables {#ff}

What were the dominant ideas at the genesis of U.S. social research? How would we know them if we found them? I consider two sets of approaches, the first methodological (how?), and the second ontological (what, why?), which will help orient us to the plan of the study. I will briefly describe a version of the familiar distinctions among qualitative and quantitative methods on one hand and nomathetic and idiographic ontologies on the other. What is new is the question of whether computational text analysis may blur the lines between these classic social science epistemologies. In asnwering this question I will organize some existing literature into four categories of research.

## Method *or* The Tortoise and the Hare

Methods are procedures for arriving at results, and they act by exposing assumptions to empirical observations. They provide a source of influence on arguments that breaks the circularity of their reasoning. Methods differ in the amount of material exposed to the observer and the speed of its exposure. The difference between human and machine learning concerns both, but speed is especially salient. 

> "You may deride my awkward pace,  
> But *slow and steady* wins the race."  
> -- Robert Lloyd [-@Lloyd1762Poems\:38]

Lloyd's version of Aesop's contestants correctly describe the consequences of haste.

First consider a humanist, a historian, a scholar with sense enough to read primary source material. How would she proceed to conduct an intellectual history? Slow and steady, a tortoise would identify diverse documentary sources allowing her to collect the names of important people and organizations and learn what she could of their biographical facts and event timelines. She would then read the scholarship both produced and consumed by these important actors. The identification of ideas would be the most difficult of her tasks, not just because reading takes time and effort (and there would be much of it), but because ideas exist in the minds of people and leave no direct empirical trace. The interpretation of writing in a historical context would suggest candidate ideas; their prevalence across time and place would indicate their historical importance. It would be tedious work requiring intelligence and patience.

Next consider a sociologist, a computer scientist, a librarian awash in books with no time to read them. She would identify a convenient source and ask a computer to do the rest. Hasty, impudent, and lacking the tortoise’s fortitude and patience, a hare uses a mental prosthetic to achieve and perhaps exceed the scale at which a humanist can consume documentation. She learns much less than the tortoise because, whereas people can find what they were not looking for, machines can learn only what they are told to learn. She fails in the test of knowledge (of course we know who wins in the end), but at least she fails quickly.

For shorthand, we can refer to these two approaches are Aesop’s tortoise and hare, the humanist mode and the computational mode.^[The tortoise may just as well be a hammer, and the hare a steam drill, but for the fact that the hammer ended up losing his contest.]

## Ontology *or* The Fox and the Hedgehog 

So we might learn more slowly or less quickly, but what will we be trying to learn, and why? These concerns the question of ontology. Ontologies are categories of being, or more simply, they are the assumptions that answer the (usually implicit) question, "What is this?". Ontologies may be descriptive categories of classification or explanatory causal mechanisms. It is neither possible to describe nor explain without making ontological assumptions.

> "The fox knows many things,  
> but the hedgehog knows one big thing."  
> -- Isaiah Berlin[-@Berlin1953Hedgehog\:1]

Here two of Berlin’s creatures will help show two different ontological approaches.

Consider a surveyor, an ethnographer, a data scientist. A fox believes that the world is nothing but the facts about it. She sets out to learn something about everything, and in so doing she tends to locate where the action happens to be without having known it was there to begin with. Though she can only skim for surface features, a fox’s shallower understanding of many things is usually very helpful. A fox learns where the important, useful, or interesting things in the world are hidden.

Next consider a theorist, a statistician, a case worker. A hedgehog learns something, but not necessarily everything, about something. They know less than a fox, but if they are a good hedgehog then what they know is good enough. A hedgehog experiences the world in filtered fashion, bothering to remember only what contributes to her system or her obsession.


So let us run Aesop’s race again, but make it a three-legged race competing in teams. And to add some purpose above crossing the finish line, the party set an objective. The teams must gather fruit to bring back to the table for supper.  Who would win? Hare and hedgehog of course, for they knew a big melon would provide more than their share, which they found at a farm stand down the road. The real contest was for second place. Hare and fox set out at once to forage, darting furiously here and there to gather a great assortment of wild nuts and berries. There were many mouths to feed and each morsel was small, so they had to be dedicated to growing a larder. Tortoise and hedgehog heard tale of a durian, the spiked and malodorous king of fruits, and thought it fit well with hedgehog’s general motif; but they did not sell them at the country market so they walked into town looking for an importer. They returned at sunset, durian in tow, to find hare and fox laying exhausted atop an impressive heap of produce that hare and hedgehog had clearly been snacking on while they idled the afternoon away. Sadly tortoise and hare never even placed; they had thought it prudent to plant a garden of their own and left for the seed store with no intention of returning to the party. In a few years time they intended to be the life of the potlach.
<!-- ^[Tortoise and hare had recently fallen out with ant and grasshopper, requiring them to find new friends.] -->

## Plan of the Study *or* A Three-Legged Table

So in four corners of a table, three were full of different produce and in one lay only a promise.

```{r fab-tab, include=T}
fab<-data.frame(matrix(rep(NA_character_,4),ncol=2),stringsAsFactors = F)
colnames(fab)<-ec('fox,hedgehog')
rownames(fab)<-ec('tortoise,hare')
fab['tortoise','fox']='Historian, gold standard'
fab['tortoise','hedgehog']='Weber & Marx'
fab['hare','fox']='Historian, gold standard'
fab['tortoise','hedgehog']='Historian, gold standard'
sg(fab %>% data.table(keep.rownames = T) %>% setnames('rn',' '),tit="Fable Table") 
```


> "One’s own consciousness dances about upon the words like a will-o’-the-wisp."  
> --Niklas Luhmann [-@Luhmann2002Theories\:166]

```{r dgr-ont,include=T,fig.cap='Ontological Approaches to Topics',screenshot.force = T}
# DiagrammeR::grViz('
#   graph {
#     graph [center=true]
#     node [shape=plaintext]
#     topic--useful
#     topic--true
#     true--thought
#     true--communication
#     true--artifact
#   }
#   ') 
DiagrammeR::grViz('
graph {
  node [style=plaintext]
  subgraph cluster_1 {
    label = "phone"
    1 [shape=plaintext,label="machine"]
    subgraph cluster_2 {
      style=filled
      fillcolor=gray
      label = "pheme"
      2 [shape=plaintext,label="machine\nhuman"]
      subgraph cluster_3 {
        style=filled
        fillcolor=white
        label = "rheme"
        3 [shape=plaintext,label="human"]
      }
    }
  }
}
')
```

Text as thought, or as communication. Thought holds that the ideas can be reliably interpreted, perhaps hermeneutically, to recover the mental events or intentions of authors and readers. Ambitious. Easier is to treat texts as communications, as messages, and worry little about their meanings or interpretations. A study of communication is a good foundation for the study of thought, but it is a separate task and the one we undertake here.

## History of Ideas {#kd-problem}

Though the intellectual history of the social sciences begins well before the Civil War, the current epoch of its institutional history, the epoch of professions, becomes possible only in the postbellum period. Before the Civil War, social research was a skilled occupation, and individual researchers found patrons through government and civil institutions supporting pursuits of knowledge, such as the American Philosophical Society (APS) founded in 1743 in Philadelphia. One of the great consequences of war was renewed federal investment in nation building in its aftermath. Some Union congressmen saw the rebellion as a failure of education, prompting the creation of the U.S. Department of Education in 1867 to strengthen nation building. In the postbellum period both anthropology and sociology develop as professions due primarily to the growth of universities as a new context of their activity. Without universities, social research would have remained an occupation in need of clientele. With them, social research develops the resources enabling relative autonomy, self reproduction, and occupational closure, the hallmarks of a profession. This study seeks to examine the period of transition after the Civil War from social research as occupation to social research as profession. 

### Antebellum Social Research 1783-1865

U.S. nation building had continued since the end of the American Revolution and had enrolled researchers in the projects of westward expansion against native peoples, the consolidation of slave economies against Africans, and the legitimation of the American experiment against European detractors. These were pressing problems to the intellectuals among government leaders at different levels, and they worked to make investments in new knowledge to resolve them. Such new knowledge was initially an extension of older "theories of man" in theology and enlightenment natural philosophy. These had a foothold in the private education of the American so-called natural aristocracy as well as in urban colonial institutions like the APS, which served as meeting places for intellectual elites and scholars. After the British burned the Library of Congress in 1914 Thomas Jefferson famously sold his personal library to Congress to restore it, an illustration that secular arts and sciences were produced and maintained by and under the patronage of private elites. 

The "theory of man" in the colonial period was that historical progress played out along a scale from savagery to barbarism to civilization, and that movement in the direction of progress was a function of time and innate capacity. European societies of the Old World had the best endowment of each, for they were very old (seen as continuous with pre-Christian antiquity) and were lead by aristocracies and monarchies representing the highest human capacities. Conversely colonized peoples, and colonial societies themselves, were understood to suffer low endowments of both, developing slowly due to their lack of morality, reason, and aesthetics and having had less time to achieve what little progress they could.

To gain parity with their colonial masters, the American elite altered the theory to emphasize capacity over time. Because all men were equal, the Americas could achieve greatness meeting or surpassing Europe given independence and opportunity. Their masters scoffed at the idea that a colony could ever "catch up" to the development of Europe, with aristocratic families charting their roots into antiquity. American elites, especially Southerners like Jefferson, saw themselves as naturally superior to their countrymen, taking an aristocratic role in their own country, but not requiring the ancient lineages qualifying aristocracy in Europe. The values of the Declaration were egalitarian only in this limited sense of parity across civilizations.

Egalitarianism did not mean parity within civilizations. Colonial theologians, philosophers, and learned elites had long provided the intellectual rationalizations of the domination of women and the environment, and such reasoning was readily exported to the yet unresolved problems of native tribes and African slaves. Paradoxically egalitarian values reinforced rather than ameliorated supremacist ideologies at home. If all men were equal, and natives (or Africans, or women) were clearly inferior, then natives must not be men. The racism of the Old World order in which hierarchy and domination were natural and unproblematic, was also a consequence of egalitarian values but for different reasons. Such racism was indeed more pernicious in America as the concept of equality invited even more ardent "proof" of the inferiority of dominated peoples.

American intellectuals knew the same logic of inferiority could be applied to the American elite who if they could not match the rate of progress of European society would be judged a newly diverging and potentially inferior race of men. Though it did little to ameliorate white male supremacist thinking, the contradictions between the universalistic and egalitarian values expressed in the U.S. Constitution did occasionally create grist for academic debate leading to some investments in research critical of extant racist typologies.  In Europe scholars of comparative languages charted the supposed common origins of European cultures and the time it had taken them to diverge. This inspired Jefferson to patronize ethnological research on the languages of native tribes in America. He had by 1785 amassed enough data to appreciate their diversity, which he took as evidence that New World cultures may be older, and therefore more developed, than commonly assumed by colonizers.

The egalitarian-for-its-time thinking of Jefferson did not contravene racism, but it did provide a different model of racial hierarchy, which could be consequential for government projects. Jefferson thought that if all men are equal, that they share the same human capacity for progress, then native people had merely not had the time to develop and discover progress for themselves. They could and should be taught, especially to abandon hunting in favor of agriculture. For a brief time in the 1840s the Virginia legislature used tax incentives and educational programs to promote intermarriage between male settlers and native women with the goal of accelerating the natural development of native culture. [@Patterson2001Social\:9] Such experiments were short lived, as the predictable backlash was that such mixing would risk lowering white culture rather than elevating Indian, and genocidal and segregationist policies won out. Thus while all racial models concluded with the domination of natives by settlers, different models of racial hierarchy sewed policy disputes within colonial leadership.

The role of social research then was to be in service of either government or private political associations. Careers were made for anthropologists in patronage relationships to generate knowledge to aid nation builders. 

### Postbellum Social Research 1866-1918

While the roots of anthropology are as old as the republic, sociology did not develop in earnest until after the Civil War.
